Simplified structure
pgm: In full:
The original code generated a .json model of a data structure the loaded
the model to validate.

The new method creates a python data structure directly, then saves it
as a .json

The new method is more direct and to be preferred, but does it generate
the same results. To be validated..

 Author:    sambelltacq <Sam.Bell@d-tacq.co.uk>
 Date:      Thu Jun 8 10:32:01 2023 +0100
	modified:   HAPI/lsafhba.py